- Now the closest to the mouse position vertex is selected when selecting with S pressed. - Added images like flags. - Made thick paths.

+++ AlternativeTransport-MLaToA2018/Graph/Graph.hpp
@@ -168,7 +168,7 @@ namespace at
         void ExternalDijkstra_Unload();
         
         
-        int ZonesNum{ 0 };
+        unsigned long ZonesNum{ 0 };
         void ArcFlags_Dijkstra(Vertex* s, Vertex* t);
         void ArcFlags_Preprocessing(unsigned int mode, Vertex* vertex, Edge* edge, unsigned long index);
         double ArcFlags(unsigned long si, unsigned long ti);
